Lavet rettelser og prøvet at få vores OOP til at være overskulig Check Vigtigt
