+++ src/main/java/GameLogic.java
@@ -1,47 +1,40 @@
+import GUI_Full.GUI;
+
 import java.util.Scanner;
 
 public class GameLogic {
 
     Scanner keyboard = new Scanner(System.in);
 
+    int playerNumber = 0;
+
     public void startGame() {
 
-        int playerNumber;
+
 
         System.out.println("how many gamers are there????");
         playerNumber = keyboard.nextInt();
 
-        if (playerNumber > 1 || playerNumber < 4) {
-            GameFlow gameFlow(generatePlayers());
+
+        if (playerNumber > 1 || playerNumber <= 4) {
+            GameFlow gameFlow = new GameFlow();
+            gameFlow.generateField();
+
+
         } else {
             System.out.println("There can only be 2-4 players :'(");
             startGame();
+
         }
     }
 
-    private Player[] generatePlayers(int playerNumber) {
-        int money;
-        Player[] players = new Player[playerNumber];
+    public int getAntalPlayer()
+    {
+        return playerNumber;
+    }
 
-        if (playerNumber == 2) {
-            money = 20;
-        }
-        if (playerNumber == 3) {
-            money = 18;
-        }
-        if (playerNumber == 4) {
-            money = 16;
-        } else {
-            money = 0;
-        }
-
-        for (int i = 0; i < playerNumber; i++) {
-            System.out.println("Give me your name");
-
-            String name = keyboard.nextLine();
-
-            players[i] = new Player(name, money);
-        }
-        return players;
+    public void setAntalPlayer(int playerNumber)
+    {
+        this.playerNumber = playerNumber;
     }
 }
